# Backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "Simple TODO  API Documentation",
        "description": "Service Documentation",
        "contact": {
            "email": "info@std.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/todo/v1",
            "description": "STD API Server"
        }
    ],
    "paths": {
...
        "/category": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Authority: Users | Get all Categories",
                "description": "This endpoint retrieves all categories",
                "operationId": "categoryIndex",
                "responses": {
                    "200": {
                        "description": "Categories retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Authority: User | Create a new Category",
                "description": "This endpoint creates a new Category.",
                "operationId": "createCategory",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateCategory"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Category created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized. User not with access role"
                    },
                    "422": {
                        "description": "The given data was invalid."
                    }
                }
            }
        },
        "/category/{categoryId}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Authority: User | Get details of a Category by ID",
                "description": "This endpoint returns all related data of specified category",
                "operationId": "getCategory",
                "responses": {
                    "200": {
                        "description": "Category retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated. Token needed"
                    },
                    "403": {
                        "description": "Unauthorized. User not with access role"
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Categories"
                ],
                "summary": "Authority: User | Update a Category | Please use x-www-form-urlencoded for body",
                "description": "This endpoint updates the category",
                "operationId": "updateCategory",
                "parameters": [
                    {
                        "name": "categoryId",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCategory"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Category updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized. User not with access role"
                    },
                    "404": {
                        "description": "Category not found"
                    },
                    "422": {
                        "description": "The given data was invalid."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "Authority: User | Delete a Category",
                "description": "This endpoint deletes the selected Category",
                "operationId": "deleteCategory",
                "parameters": [
                    {
                        "name": "categoryId",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category deleted successfully"
                    },
                    "403": {
                        "description": "Unauthorized. User not with access role"
                    },
                    "404": {
                        "description": "Category not found"
                    },
                    "422": {
                        "description": "The given data was invalid."
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Category": {
                "title": "Category",
                "description": "Category model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "name",
                        "description": "The name of the Category",
                        "type": "string",
                        "example": "Laundry"
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Created at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2021-09-27 17:50:45"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "Updated at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2021-09-27 17:50:45"
                    },
                    "user_id": {
                        "title": "user_id",
                        "description": "User ID",
                        "type": "integer",
                        "example": 1
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Category"
                }
            },
...
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "name",
                        "description": "The name of the User",
                        "type": "string",
                        "example": "Peter Oduntan"
                    },
                    "email": {
                        "title": "email",
                        "description": "The email of the User",
                        "type": "string",
                        "example": "peter_oduntan@yahoo.com"
                    },
                    "thumbnail": {
                        "title": "thumbnail",
                        "description": "The image of the User",
                        "type": "string",
                        "example": "https://example.com/image.png"
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Created at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2021-09-27 17:50:45"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "Updated at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2021-09-27 17:50:45"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            },
            "CreateCategory": {
                "title": "Create Category Request",
                "description": "Create Category Request Body Data",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "The name of the category",
                        "type": "string",
                        "example": "Laundry"
                    }
                },
                "type": "object"
            },
...
            "UpdateCategory": {
                "title": "Update Category Request",
                "description": "Update Category Request Body Data",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "The name of the category",
                        "type": "string",
                        "example": "Interviews"
                    }
                },
                "type": "object"
            },
...
        }
    },
    "security": [
        []
    ]
}